pyproject.toml
[tool.poetry]
name = "evently"
version = "0.1.0"
description = "A small prototype for an event discovery web app"
authors = ["Mehdi Essalehi <mehdi.essalehi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-types = "^0.20.0"
django-stubs-ext = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
